Factor out the read path code from p4runtime_impl.cc
