+++ p4rt_app/p4runtime/BUILD.bazel
@@ -129,6 +129,7 @@ cc_library(
     deps = [
         ":ir_translation",
         ":p4info_verification",
+        ":p4runtime_read",
         ":packetio_helpers",
         ":sdn_controller_manager",
         "//gutil:io",
@@ -181,6 +182,29 @@ cc_library(
         ":p4runtime_impl",
         "@com_github_p4lang_p4runtime//:p4runtime_cc_proto",
         "@com_google_googletest//:gtest",
+    ],
+)
+
+cc_library(
+    name = "p4runtime_read",
+    srcs = ["p4runtime_read.cc"],
+    hdrs = ["p4runtime_read.h"],
+    deps = [
+        ":ir_translation",
+        "//gutil:status",
+        "//gutil:table_entry_key",
+        "//p4_pdpi:ir_cc_proto",
+        "//p4rt_app/sonic:app_db_manager",
+        "//p4rt_app/sonic:redis_connections",
+        "//p4rt_app/utils:table_utility",
+        "@boost//:bimap",
+        "@boost//:graph",
+        "@com_github_google_glog//:glog",
+        "@com_github_p4lang_p4runtime//:p4runtime_cc_proto",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings:str_format",
     ],
 )
